pyproject.toml
[tool.poetry]
name = "whimsy"
version = "0.1.0"
description = ""
authors = ["Rohan Yadav <rohanya.76426@gmail.com>"]
license = "n"
readme = "README.md"
packages = [{include = "whimsy"}]

[tool.poetry.dependencies]
python = "^3.11"
poetry = "^1.5.1"
requests = "^2.31.0"
google-generativeai = "^0.8.1"
pydantic = "^2.9.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
whimsy = "whimsy.main:main"
